+++ blocklets/ai-studio/src/components/template-form/datasets.tsx
@@ -9,7 +9,7 @@ import { useAsync } from 'react-use';
 import { TemplateYjs } from '../../../api/src/store/projects';
 import { getDatasets } from '../../libs/dataset';
 
-export default function Datasets({ form }: { form: Pick<TemplateYjs, 'datasets'> }) {
+export default function Datasets({ readOnly, form }: { readOnly?: boolean; form: Pick<TemplateYjs, 'datasets'> }) {
   const { t } = useLocaleContext();
 
   const { value: datasetsRes } = useAsync(() => getDatasets(), []);
@@ -21,6 +21,7 @@ export default function Datasets({ form }: { form: Pick<TemplateYjs, 'datasets'>
         Object.values(form.datasets).map(({ data: item }) => (
           <Box key={item.id} sx={{ display: 'flex', my: 2 }}>
             <Autocomplete
+              readOnly={readOnly}
               fullWidth
               size="small"
               value={item.vectorStore ?? null}
@@ -31,28 +32,32 @@ export default function Datasets({ form }: { form: Pick<TemplateYjs, 'datasets'>
               getOptionLabel={(v) => v.name || 'Unnamed'}
             />
 
-            <Box sx={{ display: 'flex', flexDirection: 'column', ml: 0.5 }}>
-              <Button sx={{ minWidth: 0, p: 0.2 }} onClick={() => delete form.datasets![item.id]}>
-                <Delete sx={{ fontSize: 16, color: 'grey.500' }} />
-              </Button>
-            </Box>
+            {!readOnly && (
+              <Box sx={{ display: 'flex', flexDirection: 'column', ml: 0.5 }}>
+                <Button sx={{ minWidth: 0, p: 0.2 }} onClick={() => delete form.datasets![item.id]}>
+                  <Delete sx={{ fontSize: 16, color: 'grey.500' }} />
+                </Button>
+              </Box>
+            )}
           </Box>
         ))}
 
-      <Button
-        fullWidth
-        size="small"
-        startIcon={<Add />}
-        onClick={() => {
-          const id = nanoid();
-          (getYjsValue(form) as Map<any>).doc!.transact(() => {
-            form.datasets ??= {};
-            form.datasets[id] = { index: Object.keys(form.datasets).length, data: { id, type: 'vectorStore' } };
-          });
-          setTimeout(() => document.getElementById(`option-label-${id}`)?.focus());
-        }}>
-        {t('form.add')} {t('form.dataset')}
-      </Button>
+      {!readOnly && (
+        <Button
+          fullWidth
+          size="small"
+          startIcon={<Add />}
+          onClick={() => {
+            const id = nanoid();
+            (getYjsValue(form) as Map<any>).doc!.transact(() => {
+              form.datasets ??= {};
+              form.datasets[id] = { index: Object.keys(form.datasets).length, data: { id, type: 'vectorStore' } };
+            });
+            setTimeout(() => document.getElementById(`option-label-${id}`)?.focus());
+          }}>
+          {t('form.add')} {t('form.dataset')}
+        </Button>
+      )}
     </>
   );
 }