+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.101 (2023-10-18)
+
+- fix: import lodash on demand
+- feat: simple git mode
+- fix: read only settings
+- fix: file tree scrollable & auto scroll into view
+- fix: validate debug form
+
+## 0.1.100 (2023-10-18)
+
+- chore: export the resource as project
+
 ## 0.1.99 (2023-10-17)
 
 - chore: support send ai history messages